[no-relnote] Move tools/container to cmd/nvidia-ctk-installer
