+++ cmd/nvidia-ctk-installer/container/operator/operator_test.go
