+++ cmd/nvidia-ctk-installer/container/runtime/containerd/config_v2_test.go
@@ -23,60 +23,44 @@ import (
 	testlog "github.com/sirupsen/logrus/hooks/test"
 	"github.com/stretchr/testify/require"
 
+	"github.com/NVIDIA/nvidia-container-toolkit/cmd/nvidia-ctk-installer/container"
 	"github.com/NVIDIA/nvidia-container-toolkit/pkg/config/engine/containerd"
 	"github.com/NVIDIA/nvidia-container-toolkit/pkg/config/toml"
-	"github.com/NVIDIA/nvidia-container-toolkit/tools/container"
 )
 
-func TestUpdateV1ConfigDefaultRuntime(t *testing.T) {
+const (
+	runtimeType = "runtime_type"
+)
+
+func TestUpdateV2ConfigDefaultRuntime(t *testing.T) {
 	logger, _ := testlog.NewNullLogger()
 	const runtimeDir = "/test/runtime/dir"
 
 	testCases := []struct {
-		legacyConfig                 bool
-		setAsDefault                 bool
-		runtimeName                  string
-		expectedDefaultRuntimeName   interface{}
-		expectedDefaultRuntimeBinary interface{}
+		setAsDefault               bool
+		runtimeName                string
+		expectedDefaultRuntimeName interface{}
 	}{
 		{},
 		{
-			legacyConfig:                 true,
-			setAsDefault:                 false,
-			expectedDefaultRuntimeName:   nil,
-			expectedDefaultRuntimeBinary: nil,
+			setAsDefault:               false,
+			runtimeName:                "nvidia",
+			expectedDefaultRuntimeName: nil,
 		},
 		{
-			legacyConfig:                 true,
-			setAsDefault:                 true,
-			expectedDefaultRuntimeName:   nil,
-			expectedDefaultRuntimeBinary: "/test/runtime/dir/nvidia-container-runtime",
+			setAsDefault:               false,
+			runtimeName:                "NAME",
+			expectedDefaultRuntimeName: nil,
 		},
 		{
-			legacyConfig:                 true,
-			setAsDefault:                 true,
-			runtimeName:                  "NAME",
-			expectedDefaultRuntimeName:   nil,
-			expectedDefaultRuntimeBinary: "/test/runtime/dir/nvidia-container-runtime",
+			setAsDefault:               true,
+			runtimeName:                "nvidia",
+			expectedDefaultRuntimeName: "nvidia",
 		},
 		{
-			legacyConfig:                 false,
-			setAsDefault:                 false,
-			expectedDefaultRuntimeName:   nil,
-			expectedDefaultRuntimeBinary: nil,
-		},
-		{
-			legacyConfig:                 false,
-			setAsDefault:                 true,
-			expectedDefaultRuntimeName:   "nvidia",
-			expectedDefaultRuntimeBinary: nil,
-		},
-		{
-			legacyConfig:                 false,
-			setAsDefault:                 true,
-			runtimeName:                  "NAME",
-			expectedDefaultRuntimeName:   "NAME",
-			expectedDefaultRuntimeBinary: nil,
+			setAsDefault:               true,
+			runtimeName:                "NAME",
+			expectedDefaultRuntimeName: "NAME",
 		},
 	}
 
@@ -88,41 +72,26 @@ func TestUpdateV1ConfigDefaultRuntime(t *testing.T) {
 				SetAsDefault: tc.setAsDefault,
 			}
 
-			v1, err := containerd.New(
+			v2, err := containerd.New(
 				containerd.WithLogger(logger),
 				containerd.WithConfigSource(toml.Empty),
 				containerd.WithRuntimeType(runtimeType),
-				containerd.WithUseLegacyConfig(tc.legacyConfig),
-				containerd.WithConfigVersion(1),
+				containerd.WithContainerAnnotations("cdi.k8s.io/*"),
 			)
 			require.NoError(t, err)
 
-			err = o.UpdateConfig(v1)
+			err = o.UpdateConfig(v2)
 			require.NoError(t, err)
 
-			cfg := v1.(*containerd.ConfigV1)
-			defaultRuntimeName := cfg.GetPath([]string{"plugins", "cri", "containerd", "default_runtime_name"})
+			cfg := v2.(*containerd.Config)
+
+			defaultRuntimeName := cfg.GetPath([]string{"plugins", "io.containerd.grpc.v1.cri", "containerd", "default_runtime_name"})
 			require.EqualValues(t, tc.expectedDefaultRuntimeName, defaultRuntimeName)
-
-			defaultRuntime := cfg.GetPath([]string{"plugins", "cri", "containerd", "default_runtime"})
-			if tc.expectedDefaultRuntimeBinary == nil {
-				require.Nil(t, defaultRuntime)
-			} else {
-				require.NotNil(t, defaultRuntime)
-
-				expected, err := defaultRuntimeTomlConfigV1(tc.expectedDefaultRuntimeBinary.(string))
-				require.NoError(t, err)
-
-				configContents, _ := toml.Marshal(defaultRuntime.(*toml.Tree))
-				expectedContents, _ := toml.Marshal(expected)
-
-				require.Equal(t, string(expectedContents), string(configContents), "%d: %v: %v", i, tc)
-			}
 		})
 	}
 }
 
-func TestUpdateV1Config(t *testing.T) {
+func TestUpdateV2Config(t *testing.T) {
 	logger, _ := testlog.NewNullLogger()
 	const runtimeDir = "/test/runtime/dir"
 
@@ -133,9 +102,9 @@ func TestUpdateV1Config(t *testing.T) {
 		{
 			runtimeName: "nvidia",
 			expectedConfig: map[string]interface{}{
-				"version": int64(1),
+				"version": int64(2),
 				"plugins": map[string]interface{}{
-					"cri": map[string]interface{}{
+					"io.containerd.grpc.v1.cri": map[string]interface{}{
 						"containerd": map[string]interface{}{
 							"runtimes": map[string]interface{}{
 								"nvidia": map[string]interface{}{
@@ -146,7 +115,6 @@ func TestUpdateV1Config(t *testing.T) {
 									"container_annotations":           []string{"cdi.k8s.io/*"},
 									"options": map[string]interface{}{
 										"BinaryName": "/test/runtime/dir/nvidia-container-runtime",
-										"Runtime":    "/test/runtime/dir/nvidia-container-runtime",
 									},
 								},
 								"nvidia-cdi": map[string]interface{}{
@@ -157,7 +125,6 @@ func TestUpdateV1Config(t *testing.T) {
 									"container_annotations":           []string{"cdi.k8s.io/*"},
 									"options": map[string]interface{}{
 										"BinaryName": "/test/runtime/dir/nvidia-container-runtime.cdi",
-										"Runtime":    "/test/runtime/dir/nvidia-container-runtime.cdi",
 									},
 								},
 								"nvidia-legacy": map[string]interface{}{
@@ -168,7 +135,6 @@ func TestUpdateV1Config(t *testing.T) {
 									"container_annotations":           []string{"cdi.k8s.io/*"},
 									"options": map[string]interface{}{
 										"BinaryName": "/test/runtime/dir/nvidia-container-runtime.legacy",
-										"Runtime":    "/test/runtime/dir/nvidia-container-runtime.legacy",
 									},
 								},
 							},
@@ -180,9 +146,9 @@ func TestUpdateV1Config(t *testing.T) {
 		{
 			runtimeName: "NAME",
 			expectedConfig: map[string]interface{}{
-				"version": int64(1),
+				"version": int64(2),
 				"plugins": map[string]interface{}{
-					"cri": map[string]interface{}{
+					"io.containerd.grpc.v1.cri": map[string]interface{}{
 						"containerd": map[string]interface{}{
 							"runtimes": map[string]interface{}{
 								"NAME": map[string]interface{}{
@@ -193,7 +159,6 @@ func TestUpdateV1Config(t *testing.T) {
 									"container_annotations":           []string{"cdi.k8s.io/*"},
 									"options": map[string]interface{}{
 										"BinaryName": "/test/runtime/dir/nvidia-container-runtime",
-										"Runtime":    "/test/runtime/dir/nvidia-container-runtime",
 									},
 								},
 								"nvidia-cdi": map[string]interface{}{
@@ -204,7 +169,6 @@ func TestUpdateV1Config(t *testing.T) {
 									"container_annotations":           []string{"cdi.k8s.io/*"},
 									"options": map[string]interface{}{
 										"BinaryName": "/test/runtime/dir/nvidia-container-runtime.cdi",
-										"Runtime":    "/test/runtime/dir/nvidia-container-runtime.cdi",
 									},
 								},
 								"nvidia-legacy": map[string]interface{}{
@@ -215,7 +179,6 @@ func TestUpdateV1Config(t *testing.T) {
 									"container_annotations":           []string{"cdi.k8s.io/*"},
 									"options": map[string]interface{}{
 										"BinaryName": "/test/runtime/dir/nvidia-container-runtime.legacy",
-										"Runtime":    "/test/runtime/dir/nvidia-container-runtime.legacy",
 									},
 								},
 							},
@@ -233,27 +196,27 @@ func TestUpdateV1Config(t *testing.T) {
 				RuntimeDir:  runtimeDir,
 			}
 
-			v1, err := containerd.New(
+			v2, err := containerd.New(
 				containerd.WithLogger(logger),
 				containerd.WithConfigSource(toml.Empty),
 				containerd.WithRuntimeType(runtimeType),
-				containerd.WithConfigVersion(1),
 				containerd.WithContainerAnnotations("cdi.k8s.io/*"),
 			)
 			require.NoError(t, err)
 
-			err = o.UpdateConfig(v1)
+			err = o.UpdateConfig(v2)
 			require.NoError(t, err)
 
 			expected, err := toml.TreeFromMap(tc.expectedConfig)
 			require.NoError(t, err)
 
-			require.Equal(t, expected.String(), v1.String())
+			require.Equal(t, expected.String(), v2.String())
 		})
 	}
+
 }
 
-func TestUpdateV1ConfigWithRuncPresent(t *testing.T) {
+func TestUpdateV2ConfigWithRuncPresent(t *testing.T) {
 	logger, _ := testlog.NewNullLogger()
 	const runtimeDir = "/test/runtime/dir"
 
@@ -264,9 +227,9 @@ func TestUpdateV1ConfigWithRuncPresent(t *testing.T) {
 		{
 			runtimeName: "nvidia",
 			expectedConfig: map[string]interface{}{
-				"version": int64(1),
+				"version": int64(2),
 				"plugins": map[string]interface{}{
-					"cri": map[string]interface{}{
+					"io.containerd.grpc.v1.cri": map[string]interface{}{
 						"containerd": map[string]interface{}{
 							"runtimes": map[string]interface{}{
 								"runc": map[string]interface{}{
@@ -288,7 +251,6 @@ func TestUpdateV1ConfigWithRuncPresent(t *testing.T) {
 									"options": map[string]interface{}{
 										"runc-option": "value",
 										"BinaryName":  "/test/runtime/dir/nvidia-container-runtime",
-										"Runtime":     "/test/runtime/dir/nvidia-container-runtime",
 									},
 								},
 								"nvidia-cdi": map[string]interface{}{
@@ -300,7 +262,6 @@ func TestUpdateV1ConfigWithRuncPresent(t *testing.T) {
 									"options": map[string]interface{}{
 										"runc-option": "value",
 										"BinaryName":  "/test/runtime/dir/nvidia-container-runtime.cdi",
-										"Runtime":     "/test/runtime/dir/nvidia-container-runtime.cdi",
 									},
 								},
 								"nvidia-legacy": map[string]interface{}{
@@ -312,7 +273,6 @@ func TestUpdateV1ConfigWithRuncPresent(t *testing.T) {
 									"options": map[string]interface{}{
 										"runc-option": "value",
 										"BinaryName":  "/test/runtime/dir/nvidia-container-runtime.legacy",
-										"Runtime":     "/test/runtime/dir/nvidia-container-runtime.legacy",
 									},
 								},
 							},
@@ -324,9 +284,9 @@ func TestUpdateV1ConfigWithRuncPresent(t *testing.T) {
 		{
 			runtimeName: "NAME",
 			expectedConfig: map[string]interface{}{
-				"version": int64(1),
+				"version": int64(2),
 				"plugins": map[string]interface{}{
-					"cri": map[string]interface{}{
+					"io.containerd.grpc.v1.cri": map[string]interface{}{
 						"containerd": map[string]interface{}{
 							"runtimes": map[string]interface{}{
 								"runc": map[string]interface{}{
@@ -348,7 +308,6 @@ func TestUpdateV1ConfigWithRuncPresent(t *testing.T) {
 									"options": map[string]interface{}{
 										"runc-option": "value",
 										"BinaryName":  "/test/runtime/dir/nvidia-container-runtime",
-										"Runtime":     "/test/runtime/dir/nvidia-container-runtime",
 									},
 								},
 								"nvidia-cdi": map[string]interface{}{
@@ -360,7 +319,6 @@ func TestUpdateV1ConfigWithRuncPresent(t *testing.T) {
 									"options": map[string]interface{}{
 										"runc-option": "value",
 										"BinaryName":  "/test/runtime/dir/nvidia-container-runtime.cdi",
-										"Runtime":     "/test/runtime/dir/nvidia-container-runtime.cdi",
 									},
 								},
 								"nvidia-legacy": map[string]interface{}{
@@ -372,7 +330,6 @@ func TestUpdateV1ConfigWithRuncPresent(t *testing.T) {
 									"options": map[string]interface{}{
 										"runc-option": "value",
 										"BinaryName":  "/test/runtime/dir/nvidia-container-runtime.legacy",
-										"Runtime":     "/test/runtime/dir/nvidia-container-runtime.legacy",
 									},
 								},
 							},
@@ -390,28 +347,28 @@ func TestUpdateV1ConfigWithRuncPresent(t *testing.T) {
 				RuntimeDir:  runtimeDir,
 			}
 
-			v1, err := containerd.New(
+			v2, err := containerd.New(
 				containerd.WithLogger(logger),
-				containerd.WithConfigSource(toml.FromMap(runcConfigMapV1("/runc-binary"))),
+				containerd.WithConfigSource(toml.FromMap(runcConfigMapV2("/runc-binary"))),
 				containerd.WithRuntimeType(runtimeType),
-				containerd.WithConfigVersion(1),
 				containerd.WithContainerAnnotations("cdi.k8s.io/*"),
 			)
 			require.NoError(t, err)
 
-			err = o.UpdateConfig(v1)
+			err = o.UpdateConfig(v2)
 			require.NoError(t, err)
 
 			expected, err := toml.TreeFromMap(tc.expectedConfig)
 			require.NoError(t, err)
 
-			require.Equal(t, expected.String(), v1.String())
+			require.Equal(t, expected.String(), v2.String())
 		})
 	}
 }
 
-func TestRevertV1Config(t *testing.T) {
+func TestRevertV2Config(t *testing.T) {
 	logger, _ := testlog.NewNullLogger()
+
 	testCases := []struct {
 		config map[string]interface {
 		}
@@ -420,19 +377,17 @@ func TestRevertV1Config(t *testing.T) {
 		{},
 		{
 			config: map[string]interface{}{
-				"version": int64(1),
+				"version": int64(2),
 			},
 		},
 		{
 			config: map[string]interface{}{
-				"version": int64(1),
+				"version": int64(2),
 				"plugins": map[string]interface{}{
-					"cri": map[string]interface{}{
+					"io.containerd.grpc.v1.cri": map[string]interface{}{
 						"containerd": map[string]interface{}{
 							"runtimes": map[string]interface{}{
-								"nvidia":        runtimeMapV1("/test/runtime/dir/nvidia-container-runtime"),
-								"nvidia-cdi":    runtimeMapV1("/test/runtime/dir/nvidia-container-runtime.cdi"),
-								"nvidia-legacy": runtimeMapV1("/test/runtime/dir/nvidia-container-runtime.legacy"),
+								"nvidia": runtimeMapV2("/test/runtime/dir/nvidia-container-runtime"),
 							},
 						},
 					},
@@ -441,16 +396,13 @@ func TestRevertV1Config(t *testing.T) {
 		},
 		{
 			config: map[string]interface{}{
-				"version": int64(1),
+				"version": int64(2),
 				"plugins": map[string]interface{}{
-					"cri": map[string]interface{}{
+					"io.containerd.grpc.v1.cri": map[string]interface{}{
 						"containerd": map[string]interface{}{
 							"runtimes": map[string]interface{}{
-								"nvidia":        runtimeMapV1("/test/runtime/dir/nvidia-container-runtime"),
-								"nvidia-cdi":    runtimeMapV1("/test/runtime/dir/nvidia-container-runtime.cdi"),
-								"nvidia-legacy": runtimeMapV1("/test/runtime/dir/nvidia-container-runtime.legacy"),
+								"nvidia": runtimeMapV2("/test/runtime/dir/nvidia-container-runtime"),
 							},
-							"default_runtime":      defaultRuntimeV1("/test/runtime/dir/nvidia-container-runtime"),
 							"default_runtime_name": "nvidia",
 						},
 					},
@@ -468,28 +420,23 @@ func TestRevertV1Config(t *testing.T) {
 			expected, err := toml.TreeFromMap(tc.expected)
 			require.NoError(t, err)
 
-			v1, err := containerd.New(
+			v2, err := containerd.New(
 				containerd.WithLogger(logger),
 				containerd.WithConfigSource(toml.FromMap(tc.config)),
 				containerd.WithRuntimeType(runtimeType),
-
 				containerd.WithContainerAnnotations("cdi.k8s.io/*"),
 			)
 			require.NoError(t, err)
 
-			err = o.RevertConfig(v1)
+			err = o.RevertConfig(v2)
 			require.NoError(t, err)
 
-			require.Equal(t, expected.String(), v1.String())
+			require.Equal(t, expected.String(), v2.String())
 		})
 	}
 }
 
-func defaultRuntimeTomlConfigV1(binary string) (*toml.Tree, error) {
-	return toml.TreeFromMap(defaultRuntimeV1(binary))
-}
-
-func defaultRuntimeV1(binary string) map[string]interface{} {
+func runtimeMapV2(binary string) map[string]interface{} {
 	return map[string]interface{}{
 		"runtime_type":                    runtimeType,
 		"runtime_root":                    "",
@@ -497,28 +444,15 @@ func defaultRuntimeV1(binary string) map[string]interface{} {
 		"privileged_without_host_devices": false,
 		"options": map[string]interface{}{
 			"BinaryName": binary,
-			"Runtime":    binary,
 		},
 	}
 }
 
-func runtimeMapV1(binary string) map[string]interface{} {
+func runcConfigMapV2(binary string) map[string]interface{} {
 	return map[string]interface{}{
-		"runtime_type":                    runtimeType,
-		"runtime_root":                    "",
-		"runtime_engine":                  "",
-		"privileged_without_host_devices": false,
-		"options": map[string]interface{}{
-			"BinaryName": binary,
-			"Runtime":    binary,
-		},
-	}
-}
-
-func runcConfigMapV1(binary string) map[string]interface{} {
-	return map[string]interface{}{
+		"version": 2,
 		"plugins": map[string]interface{}{
-			"cri": map[string]interface{}{
+			"io.containerd.grpc.v1.cri": map[string]interface{}{
 				"containerd": map[string]interface{}{
 					"runtimes": map[string]interface{}{
 						"runc": map[string]interface{}{
