+++ cmd/nvidia-ctk-installer/container/runtime/runtime.go
@@ -21,10 +21,10 @@ import (
 
 	"github.com/urfave/cli/v2"
 
-	"github.com/NVIDIA/nvidia-container-toolkit/tools/container"
-	"github.com/NVIDIA/nvidia-container-toolkit/tools/container/runtime/containerd"
-	"github.com/NVIDIA/nvidia-container-toolkit/tools/container/runtime/crio"
-	"github.com/NVIDIA/nvidia-container-toolkit/tools/container/runtime/docker"
+	"github.com/NVIDIA/nvidia-container-toolkit/cmd/nvidia-ctk-installer/container"
+	"github.com/NVIDIA/nvidia-container-toolkit/cmd/nvidia-ctk-installer/container/runtime/containerd"
+	"github.com/NVIDIA/nvidia-container-toolkit/cmd/nvidia-ctk-installer/container/runtime/crio"
+	"github.com/NVIDIA/nvidia-container-toolkit/cmd/nvidia-ctk-installer/container/runtime/docker"
 )
 
 const (
